# store,kv: snapshot doesn't cache the non-exists kv entries lead to poor 'insert ignore' performance