<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

The Snapshot should cache all the BatchGet() KV entries, including the non-exist ones.
This commit fixes a bug (start from https://github.com/pingcap/tidb/pull/12108) that non-exist KV entries are not cached, causing `insert ignore` to access TiKV every time and the performance is poor.

### What is changed and how it works?

The confusing API convention leads to this bug. Some places use len(value) = 0 in map[key]value to represent non-exist, some places use `_, ok := map[key]value; !ok` to represent non-exist ... and in the end,  we forget to cache the non-exist information.

After this change, we can see that the `tikvSnapshot.get` doesn't call rpc `SendRequest` anymore.

See the tracing result to verify that:

```
mysql> trace format='row' insert ignore into t values (16, 16);
+---------------------------------------------------+-----------------+------------+
| operation                                         | startTS         | duration   |
+---------------------------------------------------+-----------------+------------+
| trace                                             | 16:54:17.747308 | 2.522179ms |
|   └─session.Execute                               | 16:54:17.747320 | 2.471526ms |
|     ├─session.ParseSQL                            | 16:54:17.747341 | 53.34µs    |
|     ├─executor.Compile                            | 16:54:17.747617 | 170.713µs  |
|     │ └─session.getTxnFuture                      | 16:54:17.747644 | 4.3µs      |
|     └─session.runStmt                             | 16:54:17.747827 | 1.886016ms |
|       ├─executor.handleNoDelayExecutor            | 16:54:17.747899 | 856.666µs  |
|       │ └─*executor.InsertExec.Next               | 16:54:17.747911 | 789.649µs  |
|       │   ├─prefetchUniqueIndices                 | 16:54:17.747985 | 583.39µs   |
|       │   │ └─tikvTxn.BatchGet                    | 16:54:17.747995 | 529.101µs  |
|       │   │   └─RPCClient.SendRequest             | 16:54:17.748104 | 362.006µs  |
|       │   └─tikvSnapshot.get                      | 16:54:17.748599 | 31.144µs   |
|       └─session.CommitTxn                         | 16:54:17.748805 | 882.35µs   |
|         └─session.doCommitWitRetry                | 16:54:17.748812 | 851.158µs  |
|           └─tikvTxn.Commit                        | 16:54:17.748822 | 770.733µs  |
|             ├─twoPhaseCommitter.prewriteKeys      | 16:54:17.748846 | 297.166µs  |
|             │ └─RPCClient.SendRequest             | 16:54:17.748897 | 209.762µs  |
|             ├─tikvStore.getTimestampWithRetry     | 16:54:17.749197 | 12.927µs   |
|             └─twoPhaseCommitter.commitKeys        | 16:54:17.749260 | 273.851µs  |
|               └─RPCClient.SendRequest             | 16:54:17.749301 | 172.921µs  |
+---------------------------------------------------+-----------------+------------+
20 rows in set (0.00 sec)
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Manual test (add detailed scripts or steps below)

This is not a bug related to correctness.

Side effects

 - Possible performance boost (huge, in some cases)


Related changes

 - Need to cherry-pick to the release branch

Release note

 - Write release note for bug-fix or new feature.